feat: Enhance token fee service with admin cache management and improve documentation
