+++ src/cashier_backend/src/api/admin.rs
@@ -91,7 +91,23 @@ pub fn is_inspect_message_enabled() -> bool {
     state.settings.is_inspect_message_enabled()
 }
 
-/// Clears all cached token fees
+/// Clears all cached token fees from the service.
+///
+/// This admin endpoint invalidates all cached token transfer fees, forcing
+/// subsequent fee queries to fetch fresh data from their respective token canisters.
+/// Useful for cache invalidation when fee structures change or for testing purposes.
+///
+/// # Authorization
+///
+/// Requires `Permission::Admin`. The caller must have admin permissions or the call will panic.
+///
+/// # Returns
+///
+/// Returns `Ok(())` on successful cache clearance.
+///
+/// # Errors
+///
+/// Currently always returns `Ok(())` after clearing the cache.
 #[update]
 pub fn admin_fee_cache_clear() -> Result<(), CanisterError> {
     debug!("[admin_fee_cache_clear]");
@@ -106,7 +122,27 @@ pub fn admin_fee_cache_clear() -> Result<(), CanisterError> {
     Ok(())
 }
 
-/// Clears cached fee for specific token
+/// Clears the cached fee for a specific token.
+///
+/// This admin endpoint invalidates the cached transfer fee for a single token,
+/// forcing the next fee query for that token to fetch fresh data from its canister.
+/// Useful when a specific token's fee structure changes without affecting other tokens.
+///
+/// # Arguments
+///
+/// * `token_id` - The `Principal` of the token canister whose cached fee should be cleared
+///
+/// # Authorization
+///
+/// Requires `Permission::Admin`. The caller must have admin permissions or the call will panic.
+///
+/// # Returns
+///
+/// Returns `Ok(())` on successful cache clearance for the specified token.
+///
+/// # Errors
+///
+/// Currently always returns `Ok(())` after clearing the token's cached fee.
 #[update]
 pub fn admin_fee_cache_clear_token(token_id: Principal) -> Result<(), CanisterError> {
     debug!("[admin_fee_cache_clear_token] token_id={}", token_id);
