+++ src/cashier_backend/src/api/state.rs
@@ -7,8 +7,7 @@ use crate::{
         token_fee::{IcrcTokenFetcher, TokenFeeService},
     },
     repositories::{
-        AUTH_SERVICE_STORE, LOGGER_SERVICE_STORE, ThreadlocalRepositories,
-        auth::AuthServiceStorage, token_fee::TokenFeeRepositoryStorage,
+        AUTH_SERVICE_STORE, LOGGER_SERVICE_STORE, ThreadlocalRepositories, auth::AuthServiceStorage,
     },
 };
 use cashier_common::runtime::{IcEnvironment, RealIcEnvironment};
@@ -23,8 +22,7 @@ pub struct CanisterState<E: IcEnvironment + Clone + 'static> {
     pub log_service: LoggerConfigService<&'static LocalKey<RefCell<LoggerServiceStorage>>>,
     pub request_lock_service: RequestLockService<ThreadlocalRepositories>,
     pub settings: SettingsService<ThreadlocalRepositories>,
-    pub token_fee_service:
-        TokenFeeService<&'static LocalKey<RefCell<TokenFeeRepositoryStorage>>, E, IcrcTokenFetcher>,
+    pub token_fee_service: TokenFeeService<ThreadlocalRepositories, E, IcrcTokenFetcher>,
     pub env: E,
 }
 
@@ -36,8 +34,7 @@ impl<E: IcEnvironment + Clone + 'static> CanisterState<E> {
         let transaction_manager_v2 = IcTransactionManager::new(env.clone());
         let link_v2_service = LinkV2Service::new(&*repo, Rc::new(transaction_manager_v2));
 
-        let token_fee_service =
-            TokenFeeService::from_repositories(&*repo, env.clone(), IcrcTokenFetcher::new());
+        let token_fee_service = TokenFeeService::new(&*repo, env.clone(), IcrcTokenFetcher::new());
 
         CanisterState {
             auth_service: AuthService::new(&AUTH_SERVICE_STORE),
