+++ src/cashier_backend_types/src/repository/mod.rs
@@ -12,6 +12,7 @@ pub mod link;
 pub mod link_action;
 pub mod processing_transaction;
 pub mod request_lock;
+pub mod token_fee;
 pub mod transaction;
 pub mod user_action;
 pub mod user_link;
