+++ src/cashier_backend_types/src/repository/token_fee/mod.rs
@@ -1,0 +1,10 @@
+use candid::Nat;
+
+/// Cached token fee with timestamp for TTL validation
+#[derive(Clone, Debug)]
+pub struct CachedFee {
+    /// Token transfer fee
+    pub fee: Nat,
+    /// Timestamp when cached/updated (nanoseconds since epoch)
+    pub updated_at: u64,
+}
