+++ src/transaction_manager/src/icrc_token/mod.rs
@@ -3,4 +3,3 @@
 
 pub mod service;
 pub mod types;
-pub mod utils;
